backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
pymongo>=4.6
scikit-learn>=1.4
joblib>=1.3
numpy>=1.26
pdfminer.six>=20231228
python-docx>=1.1
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.1
